Convert webdriver pytest fixtures to use pytest.fixture as a decorator.
Starting in pytest 3.7.0, calling the decorator pytest.fixture directly
and passing a function object does not work. Since this is a weird way to
set up fixtures, just migrate fixture implementations from fixtures.py
to the top-level conftest.py and use a decorator to register each fixture.

Move non-fixture helpers to tests/support/helpers.py. Move constant values
used as defaults to tests/support/defaults.py and adjust imports and uses.
Some default values are used both by fixtures in conftest.py and by helpers
in helpers.py.
