+++ webdriver/tests/support/defaults.py
@@ -1,0 +1,8 @@
+SCRIPT_TIMEOUT = 30
+PAGE_LOAD_TIMEOUT = 300
+IMPLICIT_WAIT_TIMEOUT = 0
+
+WINDOW_SIZE = (800, 600)
+
+DRIVER_HOST = '127.0.0.1'
+DRIVER_PORT = 4444
